Deleted states fixation cross and fix cross cooldown, changed the wya stimuli appears

+++ Assets/Scripts/ExperimentSetting.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class ExperimentSetting {
     public int numbOfTargets { private set; get; }
     public int numbOfDistractors { private set; get; }
-    public int depth { private set; get; }
+    public float depth { private set; get; }
     public bool targetsFarAway { private set; get; }
     public bool distractorsFarAway { private set; get; }
     public List<float> timeIntervals { private set; get; } = new List<float>() { 2.0f, 5.0f, 10.0f };
 
-    public ExperimentSetting(int NumbOftargets, int NumbOfdistractors, bool targetsFarAway, bool distractorsFarAway, int Depth, List<float> TimeIntervals) {
+    public ExperimentSetting(int NumbOftargets, int NumbOfdistractors, bool targetsFarAway, bool distractorsFarAway, float Depth, List<float> TimeIntervals) {
         this.numbOfTargets = NumbOftargets;
         this.numbOfDistractors = NumbOfdistractors;
         this.targetsFarAway = targetsFarAway;
@@ -19,7 +19,7 @@ public class ExperimentSetting {
         this.timeIntervals = TimeIntervals;
     }
 
-    public ExperimentSetting(int NumbOftargets, int NumbOfdistractors, bool targetsFarAway, bool distractorsFarAway, int Depth) {
+    public ExperimentSetting(int NumbOftargets, int NumbOfdistractors, bool targetsFarAway, bool distractorsFarAway, float Depth) {
         this.numbOfTargets = NumbOftargets;
         this.numbOfDistractors = NumbOfdistractors;
         this.targetsFarAway = targetsFarAway;
